Create Inditication of country  and service provider
This is the python code for identification of the location and service provider of any phone number.  Make the two files in as main.py and test.py.  create number variable in test.py and write the code in main.py . Run ,main.py and you will see the location and service provider of the phone number in test.py file.
Enjoy Python!!

Thanks!!